:recycle: refactor: extract SearchIndexPipeline and SearchIndexPersistence from SearchService
Reduce search.svelte.ts from 741 to 282 lines by extracting the indexing
pipeline (index/remove/clear/batch/rebuild/content-sweep/retry, 345 lines)
and IndexedDB persistence (loadIndex/saveIndex, 139 lines) into focused
collaborators. SearchService is now a thin wiring facade. All 71 tests pass.
