+++ apps/web/src/lib/services/search-index-pipeline.ts
@@ -1,0 +1,353 @@
+import type { SearchEntry } from "schema";
+import type {
+  SearchEngine,
+  ProgressiveBatchResult,
+} from "@codex/search-engine";
+import type * as Comlink from "comlink";
+import { entityDb } from "../utils/entity-db";
+import { debugStore } from "../stores/debug.svelte";
+import type {
+  SearchProgressCoordinator,
+  TimerApi,
+} from "./search-progress-coordinator";
+
+const INDEX_BATCH_SIZE = 100;
+
+type PipelineApi = Pick<
+  SearchEngine,
+  "add" | "addBatch" | "addBatchProgressive" | "remove" | "clear"
+>;
+
+export interface SearchIndexPipelineDeps {
+  db?: typeof entityDb;
+  debug?: typeof debugStore;
+  timers?: TimerApi;
+  coordinator: SearchProgressCoordinator;
+  getApi: () => Promise<Comlink.Remote<PipelineApi> | PipelineApi>;
+  onSaveRequired: (vaultId: string) => Promise<void>;
+}
+
+export class SearchIndexPipeline {
+  private db: typeof entityDb;
+  private debug: typeof debugStore;
+  private timers: TimerApi;
+  private coordinator: SearchProgressCoordinator;
+  private getApi: () => Promise<Comlink.Remote<PipelineApi> | PipelineApi>;
+  private onSaveRequired: (vaultId: string) => Promise<void>;
+  private indexQueue: Promise<void> = Promise.resolve();
+  needsFullContentSweep = false;
+
+  constructor(deps: SearchIndexPipelineDeps) {
+    this.db = deps.db ?? entityDb;
+    this.debug = deps.debug ?? debugStore;
+    this.timers = deps.timers ?? globalThis;
+    this.coordinator = deps.coordinator;
+    this.getApi = deps.getApi;
+    this.onSaveRequired = deps.onSaveRequired;
+  }
+
+  async index(entry: SearchEntry): Promise<void> {
+    const api = await this.getApi();
+    const cleanEntry = $state.snapshot(entry);
+    this.indexQueue = this.indexQueue
+      .then(() => api.add(cleanEntry))
+      .catch((err) => this.debug.warn("Index error", err));
+    return this.indexQueue;
+  }
+
+  async indexEntity(entity: any): Promise<void> {
+    return this.index(this.mapToSearchEntry(entity));
+  }
+
+  async remove(id: string): Promise<void> {
+    const api = await this.getApi();
+    this.indexQueue = this.indexQueue
+      .then(() => api.remove(id))
+      .catch((err) => this.debug.warn("Index remove error", err));
+    return this.indexQueue;
+  }
+
+  async clear(): Promise<void> {
+    const api = await this.getApi();
+    this.coordinator.isDirty = false;
+    return api.clear();
+  }
+
+  async indexBatch(
+    entities: any[],
+    context?: { runId: string; vaultId: string; totalCount: number | null },
+  ) {
+    const api = await this.getApi();
+
+    const queued = this.indexQueue.then(async () => {
+      for (let i = 0; i < entities.length; i += INDEX_BATCH_SIZE) {
+        const chunkEntries: any[] = [];
+        const end = Math.min(i + INDEX_BATCH_SIZE, entities.length);
+        for (let j = i; j < end; j++) {
+          chunkEntries.push(this.mapToSearchEntry(entities[j]));
+        }
+        const cleanChunkEntries = $state.snapshot(chunkEntries);
+
+        if (context) {
+          if (!this.coordinator.isActiveRun(context.vaultId, context.runId))
+            return;
+          const result: ProgressiveBatchResult = await api.addBatchProgressive(
+            cleanChunkEntries,
+            {
+              runId: context.runId,
+              vaultId: context.vaultId,
+              batchIndex: Math.floor(i / INDEX_BATCH_SIZE),
+              indexedBefore: this.coordinator.getIndexProgress().indexedCount,
+              totalCount: context.totalCount,
+            },
+          );
+          if (this.coordinator.isActiveRun(result.vaultId, result.runId)) {
+            const indexedCount =
+              this.coordinator.getIndexProgress().indexedCount +
+              result.acceptedCount;
+            this.coordinator.emitProgress({
+              status: "partial",
+              vaultId: result.vaultId,
+              runId: result.runId,
+              indexedCount,
+              totalCount: context.totalCount,
+              isPartial: true,
+              canRetry: false,
+              message:
+                context.totalCount === null
+                  ? "Search is still indexing."
+                  : `Search is still indexing (${indexedCount}/${context.totalCount}).`,
+              error:
+                result.failedIds.length > 0
+                  ? `Failed to index ${result.failedIds.length} records.`
+                  : null,
+            });
+          }
+        } else {
+          await api.addBatch(cleanChunkEntries);
+        }
+        await this.delay(0);
+      }
+    });
+
+    this.indexQueue = queued.catch((err) => {
+      this.debug.warn("Index batch error", err);
+    });
+
+    return context ? queued : this.indexQueue;
+  }
+
+  async rebuildFromEntities(
+    vaultId: string,
+    entities: any[],
+    source: "metadata" | "retry",
+    options: { markReady?: boolean; saveOnReady?: boolean } = {},
+  ) {
+    const runId = this.coordinator.createRunId(vaultId);
+    const markReady = options.markReady ?? true;
+    const saveOnReady = options.saveOnReady ?? true;
+    this.coordinator.emitProgress({
+      status: "rebuilding",
+      vaultId,
+      runId,
+      indexedCount: 0,
+      totalCount: entities.length,
+      isPartial: true,
+      canRetry: false,
+      message:
+        source === "retry"
+          ? "Retrying search indexing."
+          : "Search is indexing.",
+      error: null,
+    });
+
+    try {
+      await this.clear();
+      this.coordinator.pendingRetryEntities = entities;
+      await this.indexBatch(entities, {
+        runId,
+        vaultId,
+        totalCount: entities.length,
+      });
+      if (!this.coordinator.isActiveRun(vaultId, runId)) return;
+      if (!markReady) {
+        this.coordinator.emitProgress({
+          status: "partial",
+          vaultId,
+          runId,
+          indexedCount: entities.length,
+          totalCount: entities.length,
+          isPartial: true,
+          canRetry: false,
+          message: "Search is still indexing full note content.",
+          error: null,
+        });
+        return;
+      }
+      this.coordinator.emitProgress({
+        status: "ready",
+        vaultId,
+        runId,
+        indexedCount: entities.length,
+        totalCount: entities.length,
+        isPartial: false,
+        canRetry: false,
+        message: "Search is ready.",
+        error: null,
+      });
+      if (saveOnReady) {
+        await this.onSaveRequired(vaultId);
+      }
+    } catch (err) {
+      this.coordinator.failIndexing(vaultId, runId, err);
+    }
+  }
+
+  async indexContentInBackground(vaultId: string) {
+    if (this.coordinator.activeVaultId !== vaultId) return;
+
+    this.debug.log(`[SearchIndexPipeline] Starting background content sync...`);
+    const start = performance.now();
+    let indexedCount = 0;
+    const runId =
+      this.coordinator.activeRunId ?? this.coordinator.createRunId(vaultId);
+    let totalCount: number | null;
+
+    try {
+      const contentQuery = this.db.entityContent
+        .where("vaultId")
+        .equals(vaultId);
+      totalCount =
+        typeof contentQuery.count === "function"
+          ? await contentQuery.count()
+          : null;
+
+      this.coordinator.emitProgress({
+        ...this.coordinator.getIndexProgress(),
+        status: "partial",
+        indexedCount: 0,
+        totalCount,
+        message: "Search is indexing note content.",
+      });
+
+      const metadatas = await this.db.graphEntities
+        .where("vaultId")
+        .equals(vaultId)
+        .toArray();
+
+      const metaMap = new Map(metadatas.map((m: any) => [m.id, m]));
+      let offset = 0;
+
+      while (true) {
+        if (this.coordinator.activeVaultId !== vaultId) {
+          this.debug.log(
+            `[SearchIndexPipeline] Background sync aborted (vault switched).`,
+          );
+          return;
+        }
+
+        const records = await this.db.entityContent
+          .where("vaultId")
+          .equals(vaultId)
+          .offset(offset)
+          .limit(INDEX_BATCH_SIZE)
+          .toArray();
+
+        if (records.length === 0) break;
+
+        const currentBatch = [];
+        for (const record of records) {
+          const metadata = metaMap.get(record.entityId);
+          if (metadata) {
+            currentBatch.push({
+              ...metadata,
+              content: record.content,
+              lore: record.lore,
+            });
+          }
+        }
+
+        if (currentBatch.length > 0) {
+          await this.indexBatch(currentBatch, { runId, vaultId, totalCount });
+          indexedCount += currentBatch.length;
+        }
+
+        if (records.length < INDEX_BATCH_SIZE) break;
+        offset += records.length;
+
+        // Stagger batches to let the main thread and IndexedDB breathe.
+        await this.delay(500);
+      }
+
+      this.debug.log(
+        `[SearchIndexPipeline] Background sync complete. Indexed ${indexedCount} content records in ${(performance.now() - start).toFixed(2)}ms`,
+      );
+      if (this.coordinator.isActiveRun(vaultId, runId)) {
+        this.coordinator.emitProgress({
+          status: "ready",
+          vaultId,
+          runId,
+          indexedCount,
+          totalCount,
+          isPartial: false,
+          canRetry: false,
+          message: "Search is ready.",
+          error: null,
+        });
+        await this.onSaveRequired(vaultId);
+      }
+    } catch (err) {
+      this.debug.warn(
+        `[SearchIndexPipeline] Background content sync failed`,
+        err,
+      );
+      this.coordinator.retryNeedsContentSweep = true;
+      this.coordinator.failIndexing(vaultId, runId, err);
+    }
+  }
+
+  async retryIndexing(): Promise<void> {
+    if (!this.coordinator.activeVaultId) return;
+    const vaultId = this.coordinator.activeVaultId;
+    const retryContentSweep = this.coordinator.retryNeedsContentSweep;
+    this.coordinator.retryNeedsContentSweep = false;
+    const entities =
+      this.coordinator.pendingRetryEntities.length > 0
+        ? this.coordinator.pendingRetryEntities
+        : await this.db.graphEntities
+            .where("vaultId")
+            .equals(vaultId)
+            .toArray();
+    await this.rebuildFromEntities(vaultId, entities, "retry", {
+      markReady: !retryContentSweep,
+      saveOnReady: !retryContentSweep,
+    });
+    if (retryContentSweep && this.coordinator.activeVaultId === vaultId) {
+      await this.indexContentInBackground(vaultId);
+    }
+  }
+
+  private mapToSearchEntry(entity: any): SearchEntry {
+    const path = entity._path?.join("/") || `${entity.id}.md`;
+    const keywords = [
+      ...(entity.labels || entity.tags || []),
+      entity.lore || "",
+      ...Object.values(entity.metadata || {}).flat(),
+    ].join(" ");
+
+    return {
+      id: entity.id,
+      title: entity.title,
+      aliases: (entity.aliases || []).join(" "),
+      content: entity.content || "",
+      type: entity.type,
+      path,
+      keywords,
+      updatedAt: Date.now(),
+    };
+  }
+
+  private delay(ms: number): Promise<void> {
+    return new Promise((resolve) => this.timers.setTimeout(resolve, ms));
+  }
+}
